cleaned up make
