--- utils/masker-README
@@ -1,9 +1,0 @@
-microSat_and_simpleRep_hg19.wham.mask.txt
-
-Contains all 17mers found in repeat masker for microsats
-and simple repeats.  Both sense and anti sense are included.
-
-Feb 4th.
-
-md5 / md5sum
-df4f79c49a56e92b658ccc0db8aee7e1
